feat(scripts-perf-test-flamegrill): extract rendering and loading logic for perf tests in separate package
