+++ apps/perf-test-react-components/package.json
@@ -17,6 +17,7 @@
     "@fluentui/scripts-webpack": "*"
   },
   "dependencies": {
+    "@fluentui/scripts-perf-test-flamegrill": "*",
     "@griffel/core": "^1.9.0",
     "@fluentui/react-avatar": "^9.4.6",
     "@fluentui/react-button": "^9.3.6",
@@ -27,8 +28,6 @@
     "@fluentui/react-spinbutton": "^9.2.6",
     "@fluentui/react-theme": "^9.1.7",
     "@microsoft/load-themed-styles": "^1.10.26",
-    "lodash": "^4.17.15",
-    "querystring": "^0.2.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "tslib": "^2.1.0"
