+++ apps/perf-test-react-components/src/app.tsx
@@ -1,0 +1,10 @@
+import { loadScenarios, render } from '@fluentui/scripts-perf-test-flamegrill';
+
+bootstrap();
+
+function bootstrap() {
+  const reqContext = require.context('./scenarios/', false, /^\.\/[^\.]*\.(j|t)sx?$/);
+  const scenarios = loadScenarios(reqContext);
+
+  render(scenarios);
+}
