+++ apps/perf-test/package.json
@@ -17,11 +17,10 @@
     "@fluentui/scripts-webpack": "*"
   },
   "dependencies": {
+    "@fluentui/scripts-perf-test-flamegrill": "*",
     "@fluentui/example-data": "^8.4.7",
     "@fluentui/react": "^8.106.10",
     "@microsoft/load-themed-styles": "^1.10.26",
-    "lodash": "^4.17.15",
-    "querystring": "^0.2.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "tslib": "^2.1.0"
