+++ scripts/perf-test-flamegrill/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@fluentui/scripts-perf-test-flamegrill",
+  "version": "0.0.1",
+  "private": true,
+  "main": "src/index.ts",
+  "scripts": {
+    "format": "prettier -w --ignore-path ../../.prettierignore .",
+    "format:check": "yarn format -c",
+    "lint": "eslint --ext .ts,.js .",
+    "test": "jest --passWithNoTests",
+    "type-check": "tsc -b tsconfig.json"
+  },
+  "dependencies": {
+    "flamegrill": "0.2.0"
+  }
+}
