+++ scripts/perf-test-flamegrill/src/index.ts
@@ -1,0 +1,2 @@
+export { render } from './renderer';
+export { loadScenarios } from './load-scenarios';
