+++ scripts/perf-test-flamegrill/tsconfig.lib.json
@@ -1,0 +1,12 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "lib": ["ES2019", "DOM"],
+    "jsx": "react",
+    "outDir": "../../dist/out-tsc",
+    "types": ["webpack-env"]
+  },
+  "exclude": ["**/*.spec.ts", "**/*.spec.tsx", "**/*.test.ts", "**/*.test.tsx"],
+  "include": ["./**/*.ts", "./**/*.tsx", "./**/*.js"]
+}
